+++ R/backend-class.R
@@ -1,8 +1,9 @@
-
 abstract_method <- function() {
-  stop("An abstract keyring method is called. This is an internal error. ",
-       "It most likely happends because of a broken keyring backend that ",
-       "does not implement all keyring functions.")
+  stop(
+    "An abstract keyring method is called. This is an internal error. ",
+    "It most likely happends because of a broken keyring backend that ",
+    "does not implement all keyring functions."
+  )
 }
 
 #' Abstract class of a minimal keyring backend
@@ -24,6 +25,7 @@ abstract_method <- function() {
 #'                keyring = NULL)
 #' delete(service, username = NULL, keyring = NULL)
 #' list(service = NULL, keyring = NULL)
+#' list_raw(service = NULL, keyring = NULL)
 #' ```
 #'
 #' What these functions do:
@@ -39,6 +41,7 @@ abstract_method <- function() {
 #'   byte sequence of a raw vector.
 #' * `delete()` remotes a keyring item.
 #' * `list()` lists keyring items.
+#' * `list_raw()` lists keyring items, also as raw vectors.
 #'
 #' The arguments:
 #' * `service` String, the name of a service. This is used to find the
@@ -67,22 +70,37 @@ backend <- R6Class(
 
     has_keyring_support = function() FALSE,
 
-    get = function(service, username = NULL, keyring = NULL)
-      abstract_method(),
+    get = function(service, username = NULL, keyring = NULL) abstract_method(),
     get_raw = function(service, username = NULL, keyring = NULL)
       charToRaw(self$get(service, username, keyring)),
-    set = function(service, username = NULL, keyring = NULL,
-                   prompt = "Password: ")
-      abstract_method(),
-    set_with_value = function(service, username = NULL, password = NULL,
-      keyring = NULL) abstract_method(),
-    set_with_raw_value = function(service, username = NULL, password = NULL,
-      keyring = NULL) self$set_with_value(service, username,
-        rawToChar(password), keyring),
+    set = function(
+      service,
+      username = NULL,
+      keyring = NULL,
+      prompt = "Password: "
+    ) abstract_method(),
+    set_with_value = function(
+      service,
+      username = NULL,
+      password = NULL,
+      keyring = NULL
+    ) abstract_method(),
+    set_with_raw_value = function(
+      service,
+      username = NULL,
+      password = NULL,
+      keyring = NULL
+    ) self$set_with_value(service, username, rawToChar(password), keyring),
     delete = function(service, username = NULL, keyring = NULL)
       abstract_method(),
     list = function(service = NULL, keyring = NULL)
       stop("Backend does not implement 'list'"),
+    list_raw = function(service = NULL, keyring = NULL) {
+      keys <- self$list(service, keyring)
+      keys$service_raw <- lapply(keys$service, charToRaw)
+      keys$username_raw <- lapply(keys$username, charToRaw)
+      keys
+    },
 
     print = function(...) {
       d <- self$docs()
@@ -95,12 +113,17 @@ backend <- R6Class(
     ## This should be 'protected', really, but not possible in R6
     confirm_delete_keyring = function(keyring) {
       if (is.null(keyring)) {
-        stop("Cannot delete the default keyring. ",
-             "You need to specify the name of the keyring explicitly.")
+        stop(
+          "Cannot delete the default keyring. ",
+          "You need to specify the name of the keyring explicitly."
+        )
       }
       list <- self$keyring_list()
-      if (keyring %in% list$keyring &&
-          list$num_secrets[match(keyring, list$keyring)] > 0) {
+      if (
+        keyring %in%
+          list$keyring &&
+          list$num_secrets[match(keyring, list$keyring)] > 0
+      ) {
         confirmation(
           "The keyring is not empty, type 'yes' to delete it",
           "yes"
@@ -116,6 +139,7 @@ backend <- R6Class(
         set_with_value = "set a key in the keyring",
         delete = "delete a key",
         list = "list keys in a keyring",
+        list_raw = "list keys in a keyring as raw vectors",
         has_keyring_support = "TRUE if multiple keyrings are supported"
       )
     }
@@ -173,17 +197,16 @@ backend_keyrings <- R6Class(
   public = list(
     has_keyring_support = function() TRUE,
 
-    get = function(service, username = NULL, keyring = NULL)
-      abstract_method(),
-    set = function(service, username = NULL, keyring = NULL)
-      abstract_method(),
-    set_with_value = function(service, username = NULL, password = NULL,
-      keyring = NULL)
-      abstract_method(),
-    delete = function(service, username = NULL)
-      abstract_method(),
-    list = function(service = NULL, keyring = NULL)
-      abstract_method(),
+    get = function(service, username = NULL, keyring = NULL) abstract_method(),
+    set = function(service, username = NULL, keyring = NULL) abstract_method(),
+    set_with_value = function(
+      service,
+      username = NULL,
+      password = NULL,
+      keyring = NULL
+    ) abstract_method(),
+    delete = function(service, username = NULL) abstract_method(),
+    list = function(service = NULL, keyring = NULL) abstract_method(),
 
     keyring_create = function(keyring, password) abstract_method(),
     keyring_list = function() abstract_method(),
@@ -196,17 +219,20 @@ backend_keyrings <- R6Class(
     keyring_set_default = function(keyring = NULL) abstract_method(),
 
     docs = function() {
-      modifyList(super$docs(), list(
-        . = "Inherit from this class for a new backend with multiple keyrings.",
-        keyring_create = "create new keyring",
-        keyring_list = "list all keyrings",
-        keyring_delete = "delete a keyring",
-        keyring_lock = "lock a keyring",
-        keyring_unlock = "unlock a keyring",
-        keyring_is_locked = "check if a keyring is locked",
-        keyring_default = "query the default keyring",
-        keyring_set_default = "set the default keyring"
-      ))
+      modifyList(
+        super$docs(),
+        list(
+          . = "Inherit from this class for a new backend with multiple keyrings.",
+          keyring_create = "create new keyring",
+          keyring_list = "list all keyrings",
+          keyring_delete = "delete a keyring",
+          keyring_lock = "lock a keyring",
+          keyring_unlock = "unlock a keyring",
+          keyring_is_locked = "check if a keyring is locked",
+          keyring_default = "query the default keyring",
+          keyring_set_default = "set the default keyring"
+        )
+      )
     }
   )
 )