+++ tests/testthat/test-macos.R
@@ -23,7 +23,7 @@ test_that("creating keychains", {
 
   keyring <- random_keyring()
   kb <- backend_macos$new()
-  
+
   ## To avoid an interactive password
   kb$.__enclos_env__$private$keyring_create_direct(keyring, "secret123!")
 
@@ -73,7 +73,6 @@ test_that("creating keychains 2", {
 })
 
 test_that("keyring file at special location", {
-
   skip_if_not_macos()
   skip_on_cran()
 
@@ -97,33 +96,29 @@ test_that("keyring file at special location", {
 })
 
 test_that("errors", {
-
   skip_if_not_macos()
   skip_on_cran()
 
   ## Non-existing keychain
-  expect_error(
-    backend_macos$new(tempfile())$list(),
-    "cannot open keychain"
-  )
+  expect_snapshot(error = TRUE, backend_macos$new(tempfile())$list())
 
   ## Getting non-existing password
-  expect_error(
-    backend_macos$new()$get(random_service(), random_username()),
-    "cannot get password"
+  expect_snapshot(
+    error = TRUE,
+    backend_macos$new()$get(random_service(), random_username())
   )
 
   ## Deleting non-existing password
-  expect_error(
-    backend_macos$new()$delete(random_service(), random_username()),
-    "cannot delete password"
+  expect_snapshot(
+    error = TRUE,
+    backend_macos$new()$delete(random_service(), random_username())
   )
 
   ## Create keychain without access to file
   kb <- backend_macos$new()
-  expect_error(
-    kb$.__enclos_env__$private$keyring_create_direct("/xxx", "secret123!"),
-    "cannot create keychain"
+  expect_snapshot(
+    error = TRUE,
+    kb$.__enclos_env__$private$keyring_create_direct("/xxx", "secret123!")
   )
 })
 
@@ -154,4 +149,59 @@ test_that("lock/unlock keyrings", {
   expect_false(list$locked[match(keyring, list$keyring)])
 
   expect_silent(kb$keyring_delete(keyring = keyring))
+})
+
+test_that("zero bytes in keys", {
+  # warning for zero bytes
+  ffun <- function(...) {
+    list(
+      c("foo", NA_character_),
+      c("bar", "baz"),
+      list(charToRaw("foo"), as.raw(c(3, 2, 1, 0))),
+      list(charToRaw("bar"), charToRaw("baz"))
+    )
+  }
+  fake(b_macos_list, ".Call", ffun)
+  fake(b_macos_list_raw, ".Call", ffun)
+
+  expect_snapshot({
+    b_macos_list(NULL, list(keyring_file = function(...) NULL))
+    b_macos_list_raw(NULL, list(keyring_file = function(...) NULL))
+  })
+
+  # ---------------------
+
+  ffun <- function(...) {
+    list(
+      c("foo", NA_character_),
+      c("bar", NA_character_),
+      list(charToRaw("foo"), as.raw(c(3, 2, 1, 0))),
+      list(charToRaw("bar"), as.raw(c(1, 2, 0, 1, 2)))
+    )
+  }
+  fake(b_macos_list, ".Call", ffun)
+  fake(b_macos_list_raw, ".Call", ffun)
+
+  expect_snapshot({
+    b_macos_list(NULL, list(keyring_file = function(...) NULL))
+    b_macos_list_raw(NULL, list(keyring_file = function(...) NULL))
+  })
+
+  # ---------------------
+
+  ffun <- function(...) {
+    list(
+      c("foo", "baz"),
+      c("bar", NA_character_),
+      list(charToRaw("foo"), charToRaw("baz")),
+      list(charToRaw("bar"), as.raw(c(3, 2, 1, 0)))
+    )
+  }
+  fake(b_macos_list, ".Call", ffun)
+  fake(b_macos_list_raw, ".Call", ffun)
+
+  expect_snapshot({
+    b_macos_list(NULL, list(keyring_file = function(...) NULL))
+    b_macos_list_raw(NULL, list(keyring_file = function(...) NULL))
+  })
 })