+++ NEWS.md
@@ -1,3 +1,15 @@
+# keyring 1.4.0
+
+* Now the "file" backend will only be selected as the default backend
+  (via `default_backend()`) if the system keyring exists for this backend.
+  If you want to use the "file" backend without a system keyring, then
+  you'll need to select it explicitly. See `?default_backend`.
+
+* keyring now does not depend on the assertthat, openssl, rappdirs and
+  sodium packages.
+
+* New `key_list_raw()` method to return keys as raw vectors (#159).
+
 # keyring 1.3.2
 
 * keyring uses safer `*printf()` format strings (Secret Service backend).